+++ lib/simmer/suite.rb
@@ -7,6 +7,7 @@
 # LICENSE file in the root directory of this source tree.
 #
 
+require_relative 'suite/output_router'
 require_relative 'suite/pdi_output_writer'
 require_relative 'suite/results_writer'
 require_relative 'suite/result'